* external_lib/ijcp.jar: new version (1.3)
* src/ch/usi/hippoheal/healer/HealerFacade.java: fixed important and
subtle fault.  Briefrly, the instrumentation was a victim of the
dynamic checkpoint, since the various calls to collection methods
within the HealerFacade constructor were caught up in the checkpoint
and therefore restored to their initial (empty) values.  This should
be more properly dealt with within the instrumented code.  However,
this fixes the problem without too much trouble.